// apis/api-journeys-modern/src/schema/event/video/inputs/videoCompleteEventCreateInput.ts
import { builder } from '../../../builder'
import { VideoBlockSource } from '../../../enums'

export const VideoCompleteEventCreateInput = builder.inputType(
  'VideoCompleteEventCreateInput',
  {
    fields: (t) => ({
      id: t.id({ required: false }),
      blockId: t.id({ required: true }),
      stepId: t.id({ required: false }),
      position: t.float({ required: false }),
      label: t.string({ required: false }),
      value: t.field({ type: VideoBlockSource, required: false })
    })
  }
)